pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
authlib = "1.1.0"
httpx = "0.23.0"
fastapi = "0.87.0"
python-multipart = "0.0.5"
email-validator = "1.3.0"
requests = "2.28.1"
tenacity = "8.1.0"
pydantic = "1.10.2"
emails = "0.6"
pymongo = "4.3.2"
motor = "3.1.1"
pytest = "7.2.0"
aiofiles = "22.1.0"
pyjwt = "2.6.0"
starlette = "0.21.0"
fastapi-utils = "0.2.1"
itsdangerous = "2.1.2"
pydantic-choices = "0.1.1"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
gunicorn = "20.1.0"
uvicorn = "0.19.0"
flask = "2.0.3"
python-dotenv = "0.19.2"



[tool.poetry.group.dev.dependencies]
mypy = "0.991"
black = "22.10.0"
isort = "5.10.1"
autoflake = "1.7.7"
flake8 = "5.0.4"
pytest = "7.2.0"
sqlalchemy-stubs = "0.4"
pytest-cov = "4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
